python/requirements.txt
appium-python-client>=2.4.0
pytest
pytest-xdist
selenium
